fix(web): org switcher loads every page of organizations, not just the first 50
#3447 fixed the first-50 truncation on the Organizations settings page but
missed the second reader with the identical defect: orgStore.fetchOrganizations
(the org switcher). It fetched /orgs/organizations with no page/limit, the
server default returned 50 rows, and the switcher's client-side search only
filters what was loaded — so a partner with 50+ orgs could never switch to an
org past the first page, and the switcher header under-reported the org count.

fetchAllOrganizations moves from OrganizationsPage to lib/ (re-exported from
the page so its pagination tests and any importers keep working; the store
can't import a page component without a cycle) and the store now walks pages.

Found by pre-release UI QA sweep with a 126-org seed: switcher showed
'50 organizations' and search for 'QA Org 125' returned nothing.
